% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/15-fx-ontology-diff.R
\name{fx_ontology_impact_table}
\alias{fx_ontology_impact_table}
\title{Ontology change impact table}
\usage{
fx_ontology_impact_table()
}
\value{
The \verb{change_kind -> (case_impact, scoring_impact)} reference table.
}
\description{
Ontology change impact table
}
