% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-scibio.R
\name{is_asin}
\alias{is_asin}
\title{Is it an Amazon ASIN?}
\usage{
is_asin(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a \code{B0} + 8 alphanumeric ASIN.
}
\description{
Is it an Amazon ASIN?
}
\examples{
is_asin(c("B01L7DYXL0", "B08N5WRWNW", "140955525", NA))
}
\seealso{
Other scibio detectors: 
\code{\link{is_bibcode}()},
\code{\link{is_ensembl_id}()},
\code{\link{is_icd10_code}()},
\code{\link{is_icd9_code}()},
\code{\link{is_snp_id}()},
\code{\link{is_uniprot_id}()}
}
\concept{scibio detectors}
