% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-scibio.R
\name{is_bibcode}
\alias{is_bibcode}
\title{Is it an ADS bibcode?}
\usage{
is_bibcode(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for the 19-character ADS bibcode form.
}
\description{
Is it an ADS bibcode?
}
\examples{
is_bibcode(c("2004PASP..116...98K", "2003A&A...405..175M", NA))
}
\seealso{
Other scibio detectors: 
\code{\link{is_asin}()},
\code{\link{is_ensembl_id}()},
\code{\link{is_icd10_code}()},
\code{\link{is_icd9_code}()},
\code{\link{is_snp_id}()},
\code{\link{is_uniprot_id}()}
}
\concept{scibio detectors}
