% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-geography.R
\name{is_block_fips}
\alias{is_block_fips}
\title{Is it a census block FIPS code?}
\usage{
is_block_fips(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a 15-digit code with a valid state prefix.
}
\description{
Is it a census block FIPS code?
}
\examples{
is_block_fips(c("060372062001001", NA))
}
\seealso{
Other geography detectors: 
\code{\link{is_block_group_fips}()},
\code{\link{is_cbsa}()},
\code{\link{is_census_region}()},
\code{\link{is_county_fips}()},
\code{\link{is_csa}()},
\code{\link{is_metro_area_name}()},
\code{\link{is_nces_locale}()},
\code{\link{is_state_fips}()},
\code{\link{is_tract_fips}()},
\code{\link{is_urban_rural}()}
}
\concept{geography detectors}
