% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-color.R
\name{is_cmyk_color}
\alias{is_cmyk_color}
\title{Is it a CMYK quadruple?}
\usage{
is_cmyk_color(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for four comma-separated values in 0-1.
}
\description{
Is it a CMYK quadruple?
}
\examples{
is_cmyk_color(c("0.0,0.31,0.37,0.0", "1,1,1,1", "0,0,0", NA))
}
\seealso{
Other color detectors: 
\code{\link{is_hex_color}()},
\code{\link{is_hsl_color}()},
\code{\link{is_rgb_color}()}
}
\concept{color detectors}
