% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-geography.R
\name{is_csa}
\alias{is_csa}
\title{Is it a CSA code?}
\usage{
is_csa(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a 3-digit Combined Statistical Area code
in the Census delineation (\code{.csa} lookup).
}
\description{
Is it a CSA code?
}
\examples{
is_csa(c("101", "408", "999", NA))
}
\seealso{
Other geography detectors: 
\code{\link{is_block_fips}()},
\code{\link{is_block_group_fips}()},
\code{\link{is_cbsa}()},
\code{\link{is_census_region}()},
\code{\link{is_county_fips}()},
\code{\link{is_metro_area_name}()},
\code{\link{is_nces_locale}()},
\code{\link{is_state_fips}()},
\code{\link{is_tract_fips}()},
\code{\link{is_urban_rural}()}
}
\concept{geography detectors}
