% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-scibio.R
\name{is_ensembl_id}
\alias{is_ensembl_id}
\title{Is it an Ensembl gene/transcript/protein ID?}
\usage{
is_ensembl_id(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \verb{ENS[species]<G/T/P/E><11 digits>}.
}
\description{
Is it an Ensembl gene/transcript/protein ID?
}
\examples{
is_ensembl_id(c("ENSG00000139618", "ENSMUSG00000102659", NA))
}
\seealso{
Other scibio detectors: 
\code{\link{is_asin}()},
\code{\link{is_bibcode}()},
\code{\link{is_icd10_code}()},
\code{\link{is_icd9_code}()},
\code{\link{is_snp_id}()},
\code{\link{is_uniprot_id}()}
}
\concept{scibio detectors}
