% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-person.R
\name{is_first_name}
\alias{is_first_name}
\title{Is it a common US given (first) name?}
\usage{
is_first_name(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a single token matching the internal SSA
given-name list (\code{.first_names}), case- and punctuation-insensitive.
}
\description{
Is it a common US given (first) name?
}
\examples{
is_first_name(c("James", "OLIVIA", "Xqzzy", NA))
}
\seealso{
Other person detectors: 
\code{\link{is_full_name}()},
\code{\link{is_last_name}()}
}
\concept{person detectors}
