% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-color.R
\name{is_hsl_color}
\alias{is_hsl_color}
\title{Is it an HSL triple?}
\usage{
is_hsl_color(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \verb{H, S\%, L\%} with hue 0-360 and S/L 0-100\%.
}
\description{
Is it an HSL triple?
}
\examples{
is_hsl_color(c("195, 100\%, 27\%", "0, 0\%, 0\%", "195, 100, 27", NA))
}
\seealso{
Other color detectors: 
\code{\link{is_cmyk_color}()},
\code{\link{is_hex_color}()},
\code{\link{is_rgb_color}()}
}
\concept{color detectors}
