% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-scibio.R
\name{is_icd10_code}
\alias{is_icd10_code}
\title{Is it an ICD-10 code?}
\usage{
is_icd10_code(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \verb{<letter><2 digits>[.<up to 4 alnum>]}.
}
\description{
Is it an ICD-10 code?
}
\examples{
is_icd10_code(c("I61", "D76.0", "S72.001A", NA))
}
\seealso{
Other scibio detectors: 
\code{\link{is_asin}()},
\code{\link{is_bibcode}()},
\code{\link{is_ensembl_id}()},
\code{\link{is_icd9_code}()},
\code{\link{is_snp_id}()},
\code{\link{is_uniprot_id}()}
}
\concept{scibio detectors}
