% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-web.R
\name{is_ipv4}
\alias{is_ipv4}
\title{Is it an IPv4 address?}
\usage{
is_ipv4(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for four dot-separated octets in 0-255.
}
\description{
Is it an IPv4 address?
}
\examples{
is_ipv4(c("192.168.1.1", "256.1.1.1", NA))
}
\seealso{
Other web detectors: 
\code{\link{is_email}()},
\code{\link{is_ipv6}()},
\code{\link{is_mac}()},
\code{\link{is_url}()}
}
\concept{web detectors}
