% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-research-id.R
\name{is_isbn}
\alias{is_isbn}
\title{Is it an ISBN (10 or 13)?}
\usage{
is_isbn(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a valid ISBN-10 or ISBN-13 (dashes/spaces
ignored) whose check digit is correct.
}
\description{
Is it an ISBN (10 or 13)?
}
\examples{
is_isbn(c("140955525-9", "978-3-16-148410-0", "140955525-8", NA))
}
\seealso{
Other research-id detectors: 
\code{\link{is_doi}()},
\code{\link{is_issn}()},
\code{\link{is_npi}()},
\code{\link{is_orcid}()}
}
\concept{research-id detectors}
