% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_iso_date}
\alias{is_iso_date}
\title{Is it an ISO 8601 date?}
\usage{
is_iso_date(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \code{YYYY-MM-DD} with valid month/day.
}
\description{
Is it an ISO 8601 date?
}
\examples{
is_iso_date(c("2016-03-15", "2016-13-01", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date_range}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_month_year}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
