% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-person.R
\name{is_last_name}
\alias{is_last_name}
\title{Is it a common US surname (last name)?}
\usage{
is_last_name(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a single token matching the internal
Census surname list (\code{.last_names}), case- and punctuation-insensitive.
}
\description{
Is it a common US surname (last name)?
}
\examples{
is_last_name(c("Smith", "garcia", "Xqzzy", NA))
}
\seealso{
Other person detectors: 
\code{\link{is_first_name}()},
\code{\link{is_full_name}()}
}
\concept{person detectors}
