% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_month_of_year}
\alias{is_month_of_year}
\title{Is it a month-of-year name?}
\usage{
is_month_of_year(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a full or three-letter English month name
(\code{January}, \code{Jan}, ...). Numeric month positions (\code{01}-\code{12}) are
intentionally excluded as too ambiguous to detect from values alone.
}
\description{
Is it a month-of-year name?
}
\examples{
is_month_of_year(c("January", "feb", "Aug", "Smarch", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date}()},
\code{\link{is_date_range}()},
\code{\link{is_day_of_week}()},
\code{\link{is_fiscal_year}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_month_year}()},
\code{\link{is_quarter}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
