% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-phone.R
\name{is_msisdn}
\alias{is_msisdn}
\title{Is it an MSISDN (mobile subscriber number)?}
\usage{
is_msisdn(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a bare 10-15 digit MSISDN.
}
\description{
Is it an MSISDN (mobile subscriber number)?
}
\details{
Loose by construction: any 10-15 digit run matches. Best used with
a variable-name hint.
}
\examples{
is_msisdn(c("919961345678", "13109976224", "12345", NA))
}
\seealso{
Other phone detectors: 
\code{\link{is_intl_phone}()},
\code{\link{is_phone}()}
}
\concept{phone detectors}
