% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-geography.R
\name{is_nces_locale}
\alias{is_nces_locale}
\title{Is it an NCES locale code?}
\usage{
is_nces_locale(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a 2-digit NCES locale code (11-13, 21-23,
31-33, 41-43 = City/Suburb/Town/Rural x Large/Midsize/Small).
}
\description{
Is it an NCES locale code?
}
\examples{
is_nces_locale(c("11", "23", "44", NA))
}
\seealso{
Other geography detectors: 
\code{\link{is_block_fips}()},
\code{\link{is_block_group_fips}()},
\code{\link{is_cbsa}()},
\code{\link{is_census_region}()},
\code{\link{is_county_fips}()},
\code{\link{is_csa}()},
\code{\link{is_metro_area_name}()},
\code{\link{is_state_fips}()},
\code{\link{is_tract_fips}()},
\code{\link{is_urban_rural}()}
}
\concept{geography detectors}
