% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-address.R
\name{is_po_box}
\alias{is_po_box}
\title{Is it a PO box?}
\usage{
is_po_box(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for PO box variants (\verb{PO Box 12}, \verb{P.O.B. 3}).
}
\description{
Is it a PO box?
}
\examples{
is_po_box(c("PO Box 1234", "P.O. Box 56", "POB 789", NA))
}
\seealso{
Other address detectors: 
\code{\link{is_address}()},
\code{\link{is_city_name}()},
\code{\link{is_city_state}()},
\code{\link{is_state_zip}()},
\code{\link{is_street_address}()}
}
\concept{address detectors}
