% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_rfc2822_datetime}
\alias{is_rfc2822_datetime}
\title{Is it an RFC 2822 / HTTP datetime?}
\usage{
is_rfc2822_datetime(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for e.g. \verb{Fri, 06 Feb 1987 06:00:37 GMT}.
}
\description{
Is it an RFC 2822 / HTTP datetime?
}
\examples{
is_rfc2822_datetime(c("Fri, 06 Feb 1987 06:00:37 GMT", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date_range}()},
\code{\link{is_day_of_week}()},
\code{\link{is_fiscal_year}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_month_of_year}()},
\code{\link{is_month_year}()},
\code{\link{is_quarter}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
