% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-scibio.R
\name{is_snp_id}
\alias{is_snp_id}
\title{Is it a dbSNP reference SNP ID?}
\usage{
is_snp_id(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \code{rs} followed by digits.
}
\description{
Is it a dbSNP reference SNP ID?
}
\examples{
is_snp_id(c("rs3918290", "rs429358", "3918290", NA))
}
\seealso{
Other scibio detectors: 
\code{\link{is_asin}()},
\code{\link{is_bibcode}()},
\code{\link{is_ensembl_id}()},
\code{\link{is_icd10_code}()},
\code{\link{is_icd9_code}()},
\code{\link{is_uniprot_id}()}
}
\concept{scibio detectors}
