% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-gov.R
\name{is_ssn}
\alias{is_ssn}
\title{Is it a US Social Security Number?}
\usage{
is_ssn(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \code{ddd-dd-dddd} passing SSN issuance rules
(area not 000/666/900-999, group not 00, serial not 0000).
}
\description{
Is it a US Social Security Number?
}
\examples{
is_ssn(c("540-39-4268", "000-12-3456", NA))
}
\seealso{
Other gov detectors: 
\code{\link{is_chinese_resident_id}()},
\code{\link{is_dea}()},
\code{\link{is_ein}()},
\code{\link{is_naics}()},
\code{\link{is_uei}()}
}
\concept{gov detectors}
