% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-address.R
\name{is_state_zip}
\alias{is_state_zip}
\title{Is it a "State ZIP" value?}
\usage{
is_state_zip(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \verb{<state abbrev or name>[,] <ZIP>}.
}
\description{
Is it a "State ZIP" value?
}
\examples{
is_state_zip(c("IN 47401", "Indiana 47401", "CA, 90210", NA))
}
\seealso{
Other address detectors: 
\code{\link{is_address}()},
\code{\link{is_city_name}()},
\code{\link{is_city_state}()},
\code{\link{is_po_box}()},
\code{\link{is_street_address}()}
}
\concept{address detectors}
