% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-gov.R
\name{is_uei}
\alias{is_uei}
\title{Is it a SAM.gov Unique Entity Identifier (UEI)?}
\usage{
is_uei(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a 12-character alphanumeric UEI (no
letters \code{I}/\code{O}, first character not \code{0}).
}
\description{
Is it a SAM.gov Unique Entity Identifier (UEI)?
}
\examples{
is_uei(c("ABC123DEF456", "O1234567890X", NA))
}
\seealso{
Other gov detectors: 
\code{\link{is_chinese_resident_id}()},
\code{\link{is_dea}()},
\code{\link{is_ein}()},
\code{\link{is_naics}()},
\code{\link{is_ssn}()}
}
\concept{gov detectors}
