% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-codes.R
\name{is_unix_time}
\alias{is_unix_time}
\title{Is it a Unix timestamp (seconds)?}
\usage{
is_unix_time(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a 9-10 digit epoch-seconds value.
}
\description{
Is it a Unix timestamp (seconds)?
}
\details{
Inherently loose: any 9-10 digit integer in the plausible epoch
range matches. Intended as a weak signal, corrected in the DGF if wrong.
}
\examples{
is_unix_time(c("1513036980", "946684800", "12345", NA))
}
\seealso{
Other code detectors: 
\code{\link{is_ca_postal_code}()},
\code{\link{is_country_code}()},
\code{\link{is_http_status}()},
\code{\link{is_roman}()},
\code{\link{is_uk_postcode}()},
\code{\link{is_us_state}()},
\code{\link{is_zip_code}()}
}
\concept{code detectors}
