% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-geography.R
\name{is_urban_rural}
\alias{is_urban_rural}
\title{Is it an urban/rural designation?}
\usage{
is_urban_rural(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for common urban/rural designation words.
}
\description{
Is it an urban/rural designation?
}
\examples{
is_urban_rural(c("Urban", "rural", "Suburban", "banana", NA))
}
\seealso{
Other geography detectors: 
\code{\link{is_block_fips}()},
\code{\link{is_block_group_fips}()},
\code{\link{is_cbsa}()},
\code{\link{is_census_region}()},
\code{\link{is_county_fips}()},
\code{\link{is_csa}()},
\code{\link{is_metro_area_name}()},
\code{\link{is_nces_locale}()},
\code{\link{is_state_fips}()},
\code{\link{is_tract_fips}()}
}
\concept{geography detectors}
