% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-web.R
\name{is_url}
\alias{is_url}
\title{Is it a URL?}
\usage{
is_url(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for an http/https/ftp URL with a dotted host.
}
\description{
Is it a URL?
}
\examples{
is_url(c("https://example.com/p?q=1", "example.com", NA))
}
\seealso{
Other web detectors: 
\code{\link{is_email}()},
\code{\link{is_ipv4}()},
\code{\link{is_ipv6}()},
\code{\link{is_mac}()}
}
\concept{web detectors}
