% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_us_datetime}
\alias{is_us_datetime}
\title{Is it a US numeric datetime with AM/PM?}
\usage{
is_us_datetime(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \verb{M/D/YYYY h:mm[:ss] [AM/PM]} (\code{/} or \code{-}).
}
\description{
Is it a US numeric datetime with AM/PM?
}
\examples{
is_us_datetime(c("2/6/1987 6:00:37 AM", "05-13-2009 12:35 PM", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date}()},
\code{\link{is_date_range}()},
\code{\link{is_day_of_week}()},
\code{\link{is_fiscal_year}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_month_of_year}()},
\code{\link{is_month_year}()},
\code{\link{is_quarter}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
