% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-geometry.R
\name{is_wkt_geometry}
\alias{is_wkt_geometry}
\title{Is it a WKT (Well-Known Text) geometry?}
\usage{
is_wkt_geometry(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a WKT geometry literal such as
\verb{POINT (30 10)} or \code{POLYGON ((...))}, optionally with a \code{Z}/\code{M}/\code{ZM}
dimensionality tag. Matches the standard OGC geometry types.
}
\description{
Is it a WKT (Well-Known Text) geometry?
}
\examples{
is_wkt_geometry(c("POINT (30 10)", "LINESTRING (0 0, 1 1)", "banana", NA))
}
\seealso{
Other geometry detectors: 
\code{\link{is_geojson}()},
\code{\link{is_wkb}()}
}
\concept{geometry detectors}
