% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_yyyymm}
\alias{is_yyyymm}
\title{Is it a compact YYYYMM year-month?}
\usage{
is_yyyymm(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a valid \code{YYYYMM} string.
}
\description{
Six digits: a four-digit year (1900-2099) followed by a two-digit month
(01-12), e.g. a fiscal reporting period like \code{201912}. Constrained to a sane
year window and a valid month so it does not swallow arbitrary six-digit
counts.
}
\examples{
is_yyyymm(c("201912", "201301", "201913", "2019", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date}()},
\code{\link{is_date_range}()},
\code{\link{is_day_of_week}()},
\code{\link{is_fiscal_year}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_month_of_year}()},
\code{\link{is_month_year}()},
\code{\link{is_quarter}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()}
}
\concept{datetime detectors}
