% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_abbr_month_date}
\alias{is_abbr_month_date}
\title{Is it a day-month-year date with an abbreviated month?}
\usage{
is_abbr_month_date(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for e.g. \code{1-Jul-14} or \code{21/Nov/2012}.
}
\description{
Is it a day-month-year date with an abbreviated month?
}
\examples{
is_abbr_month_date(c("1-Jul-14", "21/Nov/2012", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_clock_time}()},
\code{\link{is_date_range}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_month_year}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
