% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_day_of_week}
\alias{is_day_of_week}
\title{Is it a day-of-week name?}
\usage{
is_day_of_week(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for a full or three-letter English weekday
name (\code{Monday}, \code{Mon}, ...). Numeric weekday positions are intentionally
excluded as too ambiguous to detect from values alone.
}
\description{
Is it a day-of-week name?
}
\examples{
is_day_of_week(c("Monday", "TUE", "Wed", "Someday", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date_range}()},
\code{\link{is_fiscal_year}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_month_of_year}()},
\code{\link{is_month_year}()},
\code{\link{is_quarter}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
