% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_iso_datetime}
\alias{is_iso_datetime}
\title{Is it an ISO 8601 datetime?}
\usage{
is_iso_datetime(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \code{YYYY-MM-DDThh:mm:ss} with optional
fractional seconds and \code{Z}/offset zone.
}
\description{
Is it an ISO 8601 datetime?
}
\examples{
is_iso_datetime(c("2009-06-15T13:45:30+02:00", "2009-06-15", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date_range}()},
\code{\link{is_day_of_week}()},
\code{\link{is_fiscal_year}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_month_of_year}()},
\code{\link{is_month_year}()},
\code{\link{is_quarter}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
