% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_long_date}
\alias{is_long_date}
\title{Is it a verbose English date?}
\usage{
is_long_date(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for weekday/month-name dates such as
\verb{Friday, June 16, 1967} or \verb{Mar 15 2016 10:53:29}. Requires a weekday
prefix or a year, to avoid overlapping \code{\link[=is_month_day]{is_month_day()}}.
}
\description{
Is it a verbose English date?
}
\examples{
is_long_date(c("Friday, June 16, 1967", "Mar 15 2016 10:53:29", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date}()},
\code{\link{is_date_range}()},
\code{\link{is_day_of_week}()},
\code{\link{is_fiscal_year}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_month_day}()},
\code{\link{is_month_of_year}()},
\code{\link{is_month_year}()},
\code{\link{is_quarter}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
