% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-datetime.R
\name{is_month_year}
\alias{is_month_year}
\title{Is it a Month-Year value?}
\usage{
is_month_year(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
Logical vector; \code{TRUE} for \verb{Month YYYY} or \verb{Month, YYYY}.
}
\description{
Is it a Month-Year value?
}
\examples{
is_month_year(c("February 1987", "August, 2005", NA))
}
\seealso{
Other datetime detectors: 
\code{\link{is_abbr_month_date}()},
\code{\link{is_clock_time}()},
\code{\link{is_date_range}()},
\code{\link{is_hijri_date}()},
\code{\link{is_iso_date}()},
\code{\link{is_iso_datetime}()},
\code{\link{is_long_date}()},
\code{\link{is_month_day}()},
\code{\link{is_rfc2822_datetime}()},
\code{\link{is_us_date}()},
\code{\link{is_us_datetime}()},
\code{\link{is_yyyymm}()}
}
\concept{datetime detectors}
